fix(manager): hide empty job error log
