+++ apps/manager/src/features/jobs/live-job-detail-screen.tsx
@@ -1,8 +1,8 @@
 "use client"
 
 import React, { useEffect, useMemo, useState } from "react"
-import { formatStepName } from "@/lib/workflow-steps"
 import type { JobRecord } from "@/types/job"
+import { JobErrorLogSection } from "./job-error-log-section"
 import { LiveJobDetailHeader } from "./live-job-detail-header"
 import { LiveJobStepsTable } from "./live-job-steps-table"
 import { ReviewPlayerCard } from "./review-player/review-player-card"
@@ -17,18 +17,6 @@ type ReviewContextLoadState =
 type LiveJobDetailScreenProps = {
   initialJob: JobRecord
   languageLabelsById: Record<string, string>
-}
-
-function formatDate(iso?: string): string {
-  if (!iso) return "\u2013"
-  const parsed = new Date(iso)
-  if (Number.isNaN(parsed.getTime())) return "\u2013"
-  return new Intl.DateTimeFormat(undefined, {
-    month: "short",
-    day: "numeric",
-    hour: "numeric",
-    minute: "2-digit",
-  }).format(parsed)
 }
 
 export function LiveJobDetailScreen({
@@ -106,55 +94,7 @@ export function LiveJobDetailScreen({
         onJobUpdate={setJob}
       />
 
-      <section
-        className="collection-card jobs-card jobs-error-card"
-        id="error-log"
-      >
-        <div className="jobs-card-header jobs-error-header">
-          <h3 className="jobs-section-title">Error Log</h3>
-          <span className="jobs-error-count">{job.errors.length}</span>
-        </div>
-        {job.errors.length === 0 ? (
-          <p className="small">No errors recorded.</p>
-        ) : (
-          <div className="jobs-table-wrap">
-            <table className="table jobs-table jobs-error-table">
-              <thead>
-                <tr>
-                  <th>Time</th>
-                  <th>Step</th>
-                  <th>Code</th>
-                </tr>
-              </thead>
-              <tbody>
-                {job.errors.map((error, idx) => (
-                  <React.Fragment key={`${error.at}-${idx}`}>
-                    <tr className="jobs-error-primary-row">
-                      <td>{formatDate(error.at)}</td>
-                      <td>{formatStepName(error.step)}</td>
-                      <td>
-                        {error.code ? (
-                          <code className="jobs-error-code">{error.code}</code>
-                        ) : (
-                          "\u2013"
-                        )}
-                      </td>
-                    </tr>
-                    <tr className="jobs-error-secondary-row">
-                      <td colSpan={3}>
-                        <p className="jobs-error-message">{error.message}</p>
-                        <p className="jobs-error-hint">
-                          {error.operatorHint ?? "\u2013"}
-                        </p>
-                      </td>
-                    </tr>
-                  </React.Fragment>
-                ))}
-              </tbody>
-            </table>
-          </div>
-        )}
-      </section>
+      <JobErrorLogSection errors={job.errors} />
 
       <ReviewPlayerCard job={job} reviewContext={reviewContext} />
     </>
